Use format_hex_2() in more places, instead of QString::arg()
This is more self-documenting and ensures formatting is consistently
uppercase.

+++ src/gui/pattern_editor.cpp
@@ -1857,7 +1857,7 @@ static void draw_pattern_foreground(
                         clear_subcolumn();
 
                         painter.setPen(instrument);
-                        auto s = format_hex_2(uint8_t(*row_event.instr));
+                        auto s = format_hex_2(*row_event.instr);
                         draw_cells(s, subcolumn.cell_centers());
 
                         draw_top_line(subcolumn);
@@ -1870,8 +1870,8 @@ static void draw_pattern_foreground(
 
                         painter.setPen(volume);
                         auto s = subcolumn.ncell == 2
-                            ? format_hex_2(uint8_t(*row_event.volume))
-                            : format_hex_1(uint8_t(*row_event.volume));
+                            ? format_hex_2(*row_event.volume)
+                            : format_hex_1(*row_event.volume);
                         draw_cells(s, subcolumn.cell_centers());
 
                         draw_top_line(subcolumn);
